critest: Remove dependency of source code
