--- cmd/critest/benchmark.go
@@ -1,37 +1,0 @@
-/*
-Copyright 2017 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package main
-
-import (
-	"github.com/urfave/cli"
-)
-
-var benchmarkCommand = cli.Command{
-	Name:  "benchmark",
-	Usage: "Benchmark test suite for CRI",
-	Flags: []cli.Flag{
-		cli.StringFlag{
-			Name:  "number, n",
-			Value: "5",
-			Usage: "Number of PodSandbox/container in listing benchmark test.",
-		},
-	},
-	Aliases: []string{"b"},
-	Action: func(context *cli.Context) error {
-		return runTestSuite(context, true)
-	},
-}
