+++ pkg/benchmark/benchmark.go
@@ -33,11 +33,7 @@ import (
 	. "github.com/onsi/gomega"
 )
 
-func init() {
-	framework.RegisterFlags()
-}
-
-// PerformanceTest checks configuration parameters (specified through flags) and then runs
+// TestPerformance checks configuration parameters (specified through flags) and then runs
 // benchmark tests using the Ginkgo runner.
 // If a "report directory" is specified, one or more JUnit test reports will be
 // generated in this directory.
